fixed page redirection and mobile menu
