+++ aboutme.html
@@ -2,14 +2,14 @@
 
 <html lang="en">
     <head>
-        <title>Aggelos Pipilikas | Under Construction</title>
-        <link rel="icon" href="resources/images/cross.png">
+        <title>Aggelos Pipilikas | About me</title>
+        <link rel="icon" href="resources/images/ap_logo_2.png">
         <meta name="description" content="Aggelos Pipilikas' personal site">
         <meta name="keywords" content="Aggelos Pipilikas, Evaggelos Pipilikas, personal site">
         <meta name="author" content="Aggelos Pipilikas">
         <meta charset="UTF-8">
         <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/handlebars@4.7.7/dist/handlebars.js"></script>
-        <script type="text/javascript" src="resources/scripts/mainScript.js"></script>
+        <script type="text/javascript" src="resources/scripts/menuScript.js"></script>
         <link rel="stylesheet" href="resources/styles/styles.css">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
     </head>
@@ -32,10 +32,10 @@
                 <nav>
                     <ul id="header-navigation-bar">
                         <li><a href="index.html">Home</a></li>
-                        <li><a>About me</a></li>
-                        <li><a>Projects</a></li>
+                        <li><a href="aboutme.html">About me</a></li>
+                        <li><a href="projects.html">Projects</a></li>
                         <li><a>Resume</a></li>
-                        <li><a>Contact</a></li>
+                        <li><a href="contact.html">Contact</a></li>
                     </ul>
                 </nav>
             </div>
