+++ index.html
@@ -2,14 +2,15 @@
 
 <html lang="en">
     <head>
-        <title>Aggelos Pipilikas | Under Construction</title>
-        <link rel="icon" href="resources/images/cross.png">
+        <title>Aggelos Pipilikas | Home</title>
+        <link rel="icon" href="resources/images/ap_logo_2.png">
         <meta name="description" content="Aggelos Pipilikas' personal site">
         <meta name="keywords" content="Aggelos Pipilikas, Evaggelos Pipilikas, personal site">
         <meta name="author" content="Aggelos Pipilikas">
         <meta charset="UTF-8">
         <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/handlebars@4.7.7/dist/handlebars.js"></script>
         <script type="text/javascript" src="resources/scripts/mainScript.js"></script>
+        <script type="text/javascript" src="resources/scripts/menuScript.js"></script>
         <link rel="stylesheet" href="resources/styles/styles.css">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
     </head>
@@ -32,8 +33,8 @@
                 <nav>
                     <ul id="header-navigation-bar">
                         <li><a href="index.html">Home</a></li>
-                        <li><a>About me</a></li>
-                        <li><a>Projects</a></li>
+                        <li><a href="aboutme.html">About me</a></li>
+                        <li><a href="projects.html">Projects</a></li>
                         <li><a>Resume</a></li>
                         <li><a href="contact.html">Contact</a></li>
                     </ul>
@@ -43,7 +44,7 @@
 
         <main id="home-page">
             <section id="welcome-introduction-section">
-                <video id="mobile-video" src="resources/images/video-mobile-trim.mp4" muted loop autoplay></video>
+                <video playsinline id="mobile-video" src="resources/images/video-mobile-trim.mp4" muted loop autoplay></video>
                 <video id="tablet-desktop-video" src="resources/images/video.mp4" muted loop autoplay></video>
                 <div class="content">
                     <div id="window">
@@ -180,16 +181,6 @@
             <section id="news-section">
                 <h2>News</h2>
                 <div class="content">
-                    <article class="news">
-                        <h3>Welcome</h3>
-                        <p class="date">10/1/2022</p>
-                        <p class="description">Welcome to my personal website! This website has been created by me from scratch and its purpose is mainly informational. Make sure to explore among the pages. I am sure you will find all the information you need there. Also, keep an eye here in news section, since I will post everything "new" I do, in terms of new projects, update on projects etc.</p>
-                    </article>
-                    <article class="news">
-                        <h3>Welcome</h3>
-                        <p class="date">10/1/2022</p>
-                        <p class="description">Welcome to my personal website! This website has been created by me from scratch and its purpose is mainly informational. Make sure to explore among the pages. I am sure you will find all the information you need there. Also, keep an eye here in news section, since I will post everything "new" I do, in terms of new projects, update on projects etc.</p>
-                    </article>
                     <article class="news">
                         <h3>Welcome</h3>
                         <p class="date">10/1/2022</p>
