+++ projects.html
@@ -2,14 +2,15 @@
 
 <html lang="en">
     <head>
-        <title>Aggelos Pipilikas | Under Construction</title>
-        <link rel="icon" href="resources/images/ap_logo.png">
+        <title>Aggelos Pipilikas | Projects</title>
+        <link rel="icon" href="resources/images/ap_logo_2.png">
         <meta name="description" content="Aggelos Pipilikas' personal site">
         <meta name="keywords" content="Aggelos Pipilikas, Evaggelos Pipilikas, personal site">
         <meta name="author" content="Aggelos Pipilikas">
         <meta charset="UTF-8">
         <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/handlebars@4.7.7/dist/handlebars.js"></script>
-        <script type="module" src="resources/scripts/projectsScript.js"></script>
+        <script type="text/javascript" src="resources/scripts/projectsScript.js"></script>
+        <script type="text/javascript" src="resources/scripts/menuScript.js"></script>
         <link rel="stylesheet" href="resources/styles/styles.css">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
     </head>
